+++ pyDeltaRCM/sed_tools.py
@@ -1,8 +1,22 @@
 ##  tools for sediment routing algorithms and deposition/erosion
 ##  refactored by eab
 ##  sep 2019
+from math import floor, sqrt, pi
+import numpy as np
+from random import shuffle
+import matplotlib
+from matplotlib import pyplot as plt
+from scipy import ndimage
+import sys, os, re, string
+from netCDF4 import Dataset
+import time as time_lib
+from scipy.sparse import lil_matrix, csc_matrix, hstack
+import logging
+import time
 
-class sed_tools(object):
+from .shared_tools import shared_tools
+
+class sed_tools(shared_tools):
     
     def sed_route(self):
         '''route all sediment'''
